added vim macro points awarding with accumulated scaling
