+++ vifrost/src/pages/GamePage.tsx
@@ -1,10 +1,9 @@
-import { useState, useMemo, useEffect, useRef, useCallback } from "react";
+import { useState, useEffect, useCallback } from "react";
 import { useLocation, Navigate } from "react-router-dom";
 import { useOutletContext } from "react-router-dom";
-import { EditorView } from "@codemirror/view";
-import type { ViewUpdate } from "@codemirror/view";
 import { GameScreen } from "../components/GameScreen";
 import { Avatar } from "../components/Avatar";
+import { useKeybindListener } from "../hooks/useKeybindListener";
 import type {
   GameStartPayload,
   ScoreUpdateServerPayload,
@@ -12,15 +11,6 @@ import type {
 } from "../hooks/useWebSocket";
 import type { AppOutletContext } from "../App";
 import "./GamePage.css";
-
-const SCORE_NORMAL_MODE_EDIT = -5;
-const SCORE_NAV_SHORTCUT = 25; // net +20 due to movement penalty
-
-type VimModeChangeEvent = { mode: string };
-type VimCm = {
-  on(event: "vim-mode-change", h: (e: VimModeChangeEvent) => void): void;
-};
-type EditorViewWithVim = EditorView & { cm?: VimCm };
 
 export function GamePage() {
   const location = useLocation();
@@ -34,72 +24,7 @@ export function GamePage() {
   const [runResults, setRunResults] = useState<boolean[] | null>(null);
   const [isRunning, setIsRunning] = useState(false);
 
-  const sendScoreUpdateRef = useRef(sendScoreUpdate);
-  sendScoreUpdateRef.current = sendScoreUpdate;
-  const vimModeRef = useRef<string>("normal");
-  const keyBufferRef = useRef<string>("");
-  // below's currently used for comparing cursor positions
-  // post-vim keybind. e.g. "fx" to find x, if cursor actually moved
-  // then +20 net points are awarded. otherwise, no movement, no points.
-  const fNavPendingRef = useRef<number | null>(null);
-
-  const attachVimModeListener = useCallback((view: EditorViewWithVim) => {
-    view.cm?.on("vim-mode-change", (e) => {
-      vimModeRef.current = e.mode;
-      keyBufferRef.current = "";
-      fNavPendingRef.current = null;
-    });
-
-    view.dom.addEventListener(
-      "keydown",
-      (event: KeyboardEvent) => {
-        if (vimModeRef.current !== "normal") return;
-
-        // key vs. buffer (buf)
-        // buf -> what's actually been "pressed so far"
-        // key -> current key pressed
-        const key = event.key;
-        const buf = keyBufferRef.current;
-
-        if (buf === "f") {
-          if (key.length === 1) {
-            // +20 net points awarded through useMemo event listener down below
-            fNavPendingRef.current = view.state.selection.main.head;
-          }
-          keyBufferRef.current = "";
-          return;
-        }
-
-        fNavPendingRef.current = null;
-
-        if (key === "f" && buf === "") {
-          keyBufferRef.current = "f";
-          return;
-        }
-
-        if (key === "w" || key === "b") {
-          sendScoreUpdateRef.current(SCORE_NAV_SHORTCUT);
-          keyBufferRef.current = "";
-          return;
-        }
-
-        if (/^\d$/.test(key)) {
-          keyBufferRef.current = buf + key;
-          return;
-        }
-
-        // {n}j — only awards when a numeric count is present
-        if (key === "j" && /^\d+$/.test(buf)) {
-          sendScoreUpdateRef.current(SCORE_NAV_SHORTCUT);
-          keyBufferRef.current = "";
-          return;
-        }
-
-        keyBufferRef.current = "";
-      },
-      { capture: true },
-    );
-  }, []);
+  const { attachVimModeListener, scoreExtension } = useKeybindListener(sendScoreUpdate);
 
   useEffect(() => {
     if (lastMessage?.type === "score_update") {
@@ -112,29 +37,6 @@ export function GamePage() {
       setIsRunning(false);
     }
   }, [lastMessage]);
-
-  const onEditorUpdate = useCallback((update: ViewUpdate) => {
-    if (vimModeRef.current === "insert") return;
-
-    // f{char} verification: cursor must have moved for the point to count
-    if (fNavPendingRef.current !== null) {
-      const prevPos = fNavPendingRef.current;
-      fNavPendingRef.current = null;
-      if (update.state.selection.main.head !== prevPos) {
-        sendScoreUpdateRef.current(SCORE_NAV_SHORTCUT);
-      }
-      return;
-    }
-
-    if (update.selectionSet || update.docChanged) {
-      sendScoreUpdateRef.current(SCORE_NORMAL_MODE_EDIT);
-    }
-  }, []);
-
-  const scoreExtension = useMemo(
-    () => [EditorView.updateListener.of(onEditorUpdate)],
-    [onEditorUpdate],
-  );
 
   const handleRun = useCallback(() => {
     setIsRunning(true);
